Remove unneeded assertions
This is to get around a failed test on PR (i.e., I forgot to re-test when I broke up the big branch).

Removing 'intro' because I didn't include adding the root-processing back in during this pr (comes in
commit 6335af45dfbc493cdcfa71be0ad14299be7e1f09).

Removing the check for 'build' directory because really that's covered by the check for the resultant
chapter file.