Remove duplicate IDs programmatically
This commit adds a function to add a random string of numbers to
subsequent instances of IDs found in the book project; any xrefs to
those duplicate IDs are updated on a per-file (chapter) basis; if these
turn out to be incorrect they'll need to be adjusted manually by the
author (this would be expected functionality based on our current
production process).

To do this work, I also removed the code that was doing this only for
"summary" IDs, since that's now redundant.

Also note that we don't need to do this for "parts," since those are
bare files so far as Jupyter Book is concerned.

Also includes changes to make existing tests pass and add this function
to the main loop of the program.

Revision: improve tests