Add data-type and data-code-language to pre blocks
This commit adds ability to facilitate syntax highlighting in converted
Jupyter books by checking for the "highlight" div and applying the
appropriate `data-type` to its child `<pre>` tag. Information about the
_kind_ of "highlight" is included in the highlight's parent classlist,
so we are also using that information to determine if it's a python
block (which, for the purposes of this initial pass, we are targeting
support for), and if so, apply the `data-code-language` so our
highlighter can pick it up when running the builds.

We're also getting rid of classes on the spans, since Jupyter book
already tries to highlight the blocks themselves; the `<span>` tags
don't appear to bother the highlighter so we're only removing classes,
since decomposing spans while keeping their contents (and, importantly,
their spacing) is tricky at best.

To facilitate this and future work, we also added a new file to the
example_book project to test code in a more focused way via the
`code_example_python()` fixture.