Silence `jb build` to avoid over-failing in Atlas
The underlying `jb build` script we use to generate the html files we
work with is currently "over failing" in Atlas, i.e., what should be
warnings on found but not _toc.yaml-included .ipynb files is causing
builds to fail. This is undesirable as it means PEs would have to remove
files from the actual book_src directory when doing ERs, which we don't
want for a variety of reasons.

This commit adds the `-qq` flag to _really_ shut `jb build` up. For
things like missing xref notifications and missing files, we'll rely on
a combination of build-in Python errors (e.g., FileNotFound errors when
trying to read the files into BeautifulSoup objects) and regular 'ol
Atlas missing xref build log warnings to handle these cases.

It's maybe an open question if we ought to provide custom errors on
FileNotFound instead of the bare stack trace, but in the spirit of
keeping this commit minimal I thought I'd leave the default behavior as
our fallback for now.