Add data-code-language tagging for R
This commit adds a check for the rpy2 `%%R` syntax/language flag and
adds the "r" `data-code-language` appropriately, while also removing
"ipython3" from the parent class lists (not that we use it, but just in
case that gets used in the future, since it's not really ipython3).

We also ensure that we're only tagging the R language if `%%R` appears
at the very beginning of the code cell/pre block, and since authors tend
to use comments to distinguish languages in the code (since the syntax
is common across (at least) Python and R, we also remove the `%%R` info.