Refactor text processing out of file_processing
This is to set us up for future work, but also it doesn't really make
sense to have the span movement and "cleans" done in the file processing
file, since those are really text tasks, not file tasks. This just moves
those functions into a new file and moves the tests around (and also
renames the test_chapter test file, which really should have been
renamed test_file for a while now anyway.