Refactor chapter file processing
This refactor gets us data-typing for non-chapters (e.g., prefaces, but
mostly appendixes) with subparts for free, and also cleans the code up.
We move the logic to apply the chapter-wide data-types to its own
function and allow the (new/old) process_chapter_soup function to take
either a Path (single-file chapter) or a list (a chapter with subfiles).

**NOTE**: We have to import `Union` from `typing` to support hinting for
this; Python 3.10 gives us a much cleaner `|` operator, but Atlas
currently is on 3.9.

Update: parallelism in the initial if-check in process_chapter_soup