Add handling for glossaries
This commit adds HTMLBook markup for glossaries to the main <seciton> of
the "chapter" (see file_processing.py) as well as to the definition
lists that make up the body of the glossaries themselves (see
reference_processing.py).

There's a little fancy footwork going on in the `add_glossary_datatypes`
function because of the vagaries of `bs4`. The code comments should, I
hope, clarify things, but please let me know if not.

(Tests have been added as well.)

Update: minor function rename to make it hopefully clearer that we're
adding datatypes to the glossary, not the chapter file... small thing.