Update example book with bibliography
This commit adds some placeholder bibliographical information so that we
can test on it and incorporate the jupyer book style of referencing into
our script and workflow. Since our house style (Chicago) prefers
footnote citations but allows for author-date, we're going to target
author-date as it's the closest appropriate format.

NOTE: I've included both a "bibliography" file in here for testing a
book-wide bibliography, as well as an in-chapter bibliography, since
both are valid in CMS.