Update citation handling to reflect new bibliography formats
Luckily, most of the code to handle this was already written, however we
were targeting the <=0.11 version of Jupyter Book's way of handling
cross references so now we're up-to-date. Basically we check if it's a
bracketed reference (which appears to be always and only the case for
bibliographical citations as far as I can tell) and handle it like we
used to from there.

This commit updated the test for this and the code to pass it.

Update: made the if-check for it being a citation better.