Update id moving scheme for headings/xrefs
This commit fixes a hangover-bug from the original iteration of this
script wherein we were manually applying IDs to top-level (chapter)
section headings and doing the linking that way (this was a result of
the less-good linking strategy of the earlier target version of Jupyter
Book).

This commit updates it to what it should be, namely looks for an empty
span right under a section and right next to a heading, and moves that
ID to the section so Atlas can go on to find the cross reference. Minor
test/code refactor but an important one!

Update: minor language improvement to test docstring