Add example tag and markup to code cells in notebooks
This is a first pass/draft of what it might look like to have
numbered "examples" generated from jupyter notebooks. This is probably
an atypical case but we should support it.

My current idea is that we mark the cell as an example with a cell tag
of "example" (see cell metadata in the code_py.ipynb file), and then use
a first line comment to denote the UUID, and the second line comment to
provide the caption.

We'll use the refs as usual, since they're passed through as something
like the following:

```
<span class="xref myst">hello_tim_with_output</span>
```

We can then use that to fix the cross references later (probably via a
call based on the UUID we discover in the example; that way we won't
have to worry as much about false matches).

This is all very provisional but it feels like an OK start.

Update: typo fix in code example