Add "Example" capability via an "example" cell tag
Per author request (and to continue brining jupyter book capabilities in
line with asciidoc/htmlbook capabilities in Atlas), we've added a
first-pass implementation of including numbered code examples in the
jupyter book conversion. This is done by adding a cell tag the relevant
code cell's metadata, e.g.,

```json
  {
   "cell_type": "code",
   "execution_count": 5,
   "metadata": {
    "tags": [
     "example"
    ]
   },
```

This can be done either via directly editing the .ipynb file in a text
editor or in many Jupyter GUIs (e.g., VS COde).

Once they've added the tag, they can provide a name and title for the
example via the first and second lines of the code cell in comments,
e.g.,

```python

print("Hello, world!")
```

Assuming the above, this commit adds code to apply the relevant data
types, add the heading, and apply the id for the example in question, as
well as providing a number of tests around it.

Some notes:

* We are sure not to pull comments other than the expected first two
  comments in the file, and we don't do anything if only one or the
  other comment is present.
* If there is a failure, we log it, and though we don't yet have a great
  way to pass these logs out of Atlas, issues can be discovered via running
  the script locally.
* We remove any newlines left by removing the code comments
* We do this ahead of regular code processing because we need the span
  styles present to ensure we're grabbing comments as the uuid and title
  (span classes are removed when we process code to simplify things for
  our own code highlighter).

Update: refactor get name and title from examples into separate function

Update: change target div to incorporate any output for free

We can have any number of block elements inside a cell tag, and since
the "tag_example" cell includes any output as a child, by targeting
_that_ cell to for our `data-type` and to include the title, we
basically get the inclusion of any code output for free, which is
desirable.