Improve fixture and tests for durability
While it would be nice for our tests to always pull in output data,
we're not really reconverting every time anyway, and having the fixture
explicitly state what we're testing against is both clearer and more
durable (as changes to the example book files were breaking tests that
they shouldn't be breaking). This commit provides this improvement.