Fail on unsuccessful chapter processing
This commit has the script fail (and supply the offending element) to
better align with the usual Atlas build practices (e.g., in the case of
badly-formed asciidoc). By raising an error, we can then pass that
through in the build log for some amount of visibility for authors/users
of Atlas.

Update: remove obsolete assertions