Update poetry, pyproject, and poetry.lock
Ran into an issue with dependencies and management with the "tl;dr"
being that our lockfile and poetry versioning had gotten out of sync.
The solution is to essentially rebuild the project dependencies with a
newer version of poetry (now at 1.3.2; the version we were using was
not able to resolve the dependencies for `jupyter-book`).