Fix bad return bug
Previously (due to a hangover from an earlier implementation), if we
were missing a name or title from the code block, we would return _only_
the `example_cell` in question, not the whole `chapter`. This was
causing a few odd bugs but now has been fixed, and I've added a test to
guard against this in the future.