Add handling for non-anchor xref spans
If Juptyer Book can't find a target for a given `{ref}`, it will create
a `<span>` tag with `xref` and `std-ref` classes applied to it.

Since our formal example workaround adds targets after the fact, that
means that any reference to a formal example is going to be a `<span>`
instead of an `<a>`, which would get picked up and correctly converted
by our `process_internal_refs` function (which has been renamed in this
commit, since there was a typo). This commit/PR adds code to handle
these references so authors can xref to the formal code examples, as
well as tests to ensure that the conversion happens correctly as such
and in chapter processing.