Bug fix for platform surprise class styling
One of the PEs discovered that having a `pre` class on a span inside a
`code` tag displays oddly on the platform. This is a "quick fix" to
solve for this, while the actual fix -- cleaning out all classes save
those we explicitly want -- will come as a part of TOOLSDEV-60.

Update: typo fix