Improve and fix tests
Originally there was an error in test_bibliography where the footnote
text(s) at the end were being added to the bib list, causing the "should
combine" test to fail. This looked like an issue with the footnote
processing itself, so I improved that a little bit by targeting the IDs
(instead of the dt supposed to contain the ID), and then added a better
smoke test file for future jupyter book HTML/class/ID changes to
footnote markup in the future.

This, it turns out, wasn't the issue, but rather that the duplicate
chapter file (and subsequently, duplicate IDs) in the test array was
breaking things (odd that it wasn't before, but --). So instead we add a
bibliography to another file in that "combined chapter" and rip out the
footnote stuff (since it isn't relevant to that test anyway -- I did
confirm that it didn't fail with it in, though), and then the tests pass
as expected.

All of which is to say: I'm not sure it's going to be reasonable to
handle _all_ [bibliography cases, styles, etc.,][bibs] in this script,
but we at least now have slightly better tests.

[bibs]: https://jupyterbook.org/en/stable/content/citations.html#change-the-in-line-citation-style

Update: up test coverage to 100%, checking for single-sub-bib