Convert <span>s to <code>s inside <pre> blocks when keeping highlighting
There is a discrepancy between what Jupyter Book spits out and what
Atlas expects in terms of the target tags for code highlighting; out
customization of Pygments provides <code> tags inside any <pre> tag with
the appropriate data-type and data-code-language attributes, but Jupyter
book instead gives <span> tags. Without weighing in on which is more
semantic, this quick PR is to ensure that if an author chooses the
`--keep-highlighting` option, the highlighting will actually make its
way through to the web pdf and epub.